requirements.txt
asn1crypto
certifi
cffi
chardet
cryptography
future
idna
pycparser
python-telegram-bot
requests
six
urllib3
